Use the new mode from within GB
