--- packages/editor/src/components/editor-mode/style.scss
@@ -1,102 +1,0 @@
-@use "@wordpress/base-styles/colors" as *;
-
-.editor-button {
-	display: flex;
-	align-items: center;
-	cursor: pointer;
-	background: $gray-100;
-	padding: 4px;
-	position: relative;
-	border-radius: inherit;
-
-	&:hover {
-		background: $gray-200;
-	}
-}
-
-.editor-popover.components-popover .components-popover__content {
-	padding: 0;
-	border-radius: 8px;
-	box-shadow: 0px 2px 6px rgba(0, 0, 0, 0.05);
-	border: 1px solid #ddd;
-	min-width: 312px;
-	background: #ffffff;
-}
-
-.editor-menu {
-	display: flex;
-	flex-direction: column;
-	padding: 9px;
-	gap: 2px;
-	min-width: 312px;
-	border-radius: 8px;
-}
-
-.editor-menu-item {
-	// Reset button styles
-	all: unset;
-	box-sizing: border-box;
-
-	// Layout styles
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-	padding: 8px 12px;
-	gap: 8px;
-	border-radius: 2px;
-	cursor: pointer;
-	transition: background-color 0.2s ease;
-
-	&:hover:not(.is-selected) {
-		background-color: #f0f0f0;
-	}
-
-	&.is-selected {
-		background: #3858e9;
-
-		svg {
-			color: #ffffff;
-		}
-
-		.editor-menu-item-label {
-			.editor-menu-item-title, .editor-menu-item-description {
-				color: #ffffff;
-			}
-		}
-	}
-
-	&:focus-visible {
-		outline: 2px solid var(--wp-admin-theme-color, #007cba);
-		outline-offset: -2px;
-	}
-
-	svg {
-		color: #757575;
-		fill: currentColor;
-	}
-
-	.editor-menu-item-label {
-		display: flex;
-		flex-direction: column;
-		gap: 4px;
-		flex: 1;
-
-		.editor-menu-item-title {
-			font-size: 13px;
-			line-height: 20px;
-			color: #1e1e1e;
-		}
-
-		.editor-menu-item-description {
-			font-size: 12px;
-			line-height: 16px;
-			color: #757575;
-		}
-	}
-}
-
-.editor-menu-item-content {
-	display: flex;
-	flex-direction: row;
-	gap: 8px;
-}
