+++ packages/editor/src/store/reducer.js
@@ -247,15 +247,6 @@ export function postAutosavingLock( state = {}, action ) {
 	return state;
 }
 
-export function collaboratorMode( state = 'edit', action ) {
-	switch ( action.type ) {
-		case 'SET_COLLABORATOR_MODE':
-			return action.collaboratorMode;
-	}
-
-	return state;
-}
-
 /**
  * Reducer returning the post editor setting.
  *
