+++ packages/editor/src/style.scss
@@ -8,7 +8,6 @@
 @use "./components/blog-title/style.scss" as *;
 @use "./components/document-bar/style.scss" as *;
 @use "./components/document-outline/style.scss" as *;
-@use "./components/editor-mode/style.scss" as *;
 @use "./components/document-tools/style.scss" as *;
 @use "./components/editor-interface/style.scss" as *;
 @use "./components/editor-notices/style.scss" as *;
